[PATCH 09/17] [BACKPORT] ptrace: Move setting/clearing ptrace_message into ptrace_stop
Today ptrace_message is easy to overlook as it not a core part of
ptrace_stop.  It has been overlooked so much that there are places
that set ptrace_message and don't clear it, and places that never set
it.  So if you get an unlucky sequence of events the ptracer may be
able to read a ptrace_message that does not apply to the current
ptrace stop.

Move setting of ptrace_message into ptrace_stop so that it always gets
set before the stop, and always gets cleared after the stop.  This
prevents non-sense from being reported to userspace and makes
ptrace_message more visible in the ptrace API so that kernel
developers can see it.

Backported to 4.19 for ptrace_message correctness and stability.

Link: https://marc.info/?l=linux-arch&m=164124554311501&w=2
Signed-off-by: Eric W. Biederman <ebiederm@xmission.com>
Signed-off-by: Prslc <prslc113@gmail.com>